backend/requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Cache
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Image Processing (简化版转图引擎)
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
ruff==0.2.2
